Move struct rxtx_ring to rxtx_ring.h.
